Update README.md
This is really an important issue. And it would be really nice to have the infomation here as any beginner takes this as the reference. It really wasted a lot of my time, and I feel many people willing to contribute encounter this issue.
Although it's not necessary it would be nice to have it here.